[sidekick#86] Extract BootScene as shared asset preloader
Move all asset loading (spritesheets, images, audio) and global animation
registration (VFX, animal pets) from OfficeScene into a new BootScene that
extends BaseScene. BootScene runs first, preloads everything into Phaser's
shared TextureManager, then transitions to OfficeScene on completion.

- Add BOOT to SCENE_KEYS in office-asset-keys.ts
- Create boot-scene.ts with loading screen UI, all asset loads, and anim registration
- Gut OfficeScene.preload() to no-op; remove VFX/animal anim block from create()
- Register BootScene first in OfficeGame scene array
- Add 13 unit tests validating key coverage, scene order, and transition
